requirements.txt
openai==1.75.0
pandas==2.2.3
python-dotenv==1.1.0
streamlit==1.44.1
watchdog==6.0.0